move parser for PureRecords to different file, add tests for reading json files
