+++ feos-core/tests/test_parameters1.json
@@ -1,0 +1,20 @@
+[
+    {
+        "identifier": {
+            "cas": "678-9-1"
+        },
+        "molarweight": 32.08412,
+        "model_record": {
+            "a": 0.2
+        }
+    },
+    {
+        "identifier": {
+            "cas": "123-4-5"
+        },
+        "molarweight": 16.0426,
+        "model_record": {
+            "a": 0.5
+        }
+    }
+]
