+++ feos-core/tests/test_parameters_binary.json
@@ -1,0 +1,24 @@
+[
+    {
+        "id1": {
+            "cas": "123-4-5"
+        },
+        "id2": {
+            "cas": "000-00-0"
+        },
+        "model_record": {
+            "b": 12.0
+        }
+    },
+    {
+        "id1": {
+            "cas": "123-4-5"
+        },
+        "id2": {
+            "cas": "678-9-1"
+        },
+        "model_record": {
+            "b": 12.0
+        }
+    }
+]
